Move config to a dedicated package
