+++ .gitignore
@@ -2,7 +2,7 @@
 out
 build
 generated
-config
+/config/
 
 # Temporary files
 tmp
