+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "author": "code.store",
   "license": "MIT",
+  "private": true,
   "scripts": {
     "generate-gateway-config": "json2ts -i src/types/gateway-schema.json -o ./src/types/gateway-config.ts --style.singleQuote --bannerComment \"/* eslint-disable */\"",
     "generate-mesh-source-config": "json2ts -i src/types/mesh-source-schema.json -o ./src/types/mesh-source-config.ts --style.singleQuote --bannerComment \"/* eslint-disable */\"",
@@ -14,6 +15,9 @@
     "lint": "eslint '{src,apps,libs,test}/**/*.ts' --fix",
     "test": "jest --coverage --runInBand --detectOpenHandles"
   },
+  "workspaces": [
+    "packages/config"
+  ],
   "dependencies": {
     "@graphql-mesh/config": "^0.11.5",
     "@graphql-mesh/graphql": "^0.10.2",
@@ -31,7 +35,8 @@
     "ioredis": "^4.19.4",
     "uuid": "^8.3.1",
     "winston": "^3.3.3",
-    "yaml": "^1.10.0"
+    "yaml": "^1.10.0",
+    "@graphql-portal/config": "^0.0.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.9",
