+++ packages/config/src/types/gateway-config.ts
