+++ packages/config/src/types/gateway-schema.json
