+++ packages/config/src/types/mesh-source-schema.json
