+++ src/index.ts
@@ -3,7 +3,7 @@ import cluster from 'cluster';
 import { logger } from './logger';
 import { version } from '../package.json';
 import { startServer, nodeID } from './server';
-import { loadConfig } from './gateway-config';
+import { loadGatewayConfig } from '@graphql-portal/config';
 
 function handleStopSignal(): void {
   logger.info('Stop signal received');
@@ -15,7 +15,7 @@ function handleStopSignal(): void {
 
 async function start(): Promise<void> {
   // TODO: config should be read by master and communicated to forks via RPC
-  const config = await loadConfig();
+  const config = await loadGatewayConfig();
   if (config === null) {
     logger.error('Error loading the gateway.json|yaml configuration file.');
     throw new Error('Error loading the gateway.json|yaml configuration file.');
