+++ tsconfig.json
@@ -17,7 +17,10 @@
     "noImplicitReturns": true,
     "noImplicitThis": true,
     "noUnusedLocals": true,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "paths": {
+      "@graphql-portal/config": ["packages/config/src/index.ts"]
+    }
   },
   "include": [
     "src/**/*.ts"
